moved code for individual APIs into WebApi.cs
