--- Source/RunActivity/Viewer3D/WebServices/Web/API/Template/index - ApiSample.html
@@ -1,38 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-<head>
-	<title>OPEN RAILS - Sample API</title>
-    <link rel="stylesheet" type="text/css" href="css/sample.css">
-	<script src="js/apisample.js"></script>
-</head>
-
-<body onload="setInterval (ApiSample, 500)">
-    <br><h2>Open Rails Web Server - Api Sample</h2>
-
-
-    <b>Sample String Data:</b>
-    <div id="strData"></div><br />
-    <b>Sample Integer Data:</b>
-    <div id="intData"></div><br />
-    <b>Date Data:</b>
-    <div id="dateData"></div><br />
-
-    <b>String Array Members:</b>
-    <div id="arrayMember0"></div>
-    <div id="arrayMember1"></div>
-    <div id="arrayMember2"></div>
-    <div id="arrayMember3"></div>
-    <div id="arrayMember4"></div><br />
-
-    <b>Embedded Class:</b><br />
-    <b>Embedded String Data:</b>
-    <div id="embeddedStr"></div><br />
-    <b>Embedded Integer Data:</b>
-    <div id="embeddedNumb"></div><br />
-
-
-</body>
-</html>
-
-
-
