+++ Source/RunActivity/Viewer3D/WebServices/Web/API/Template/index.html
@@ -1,0 +1,46 @@
+<!DOCTYPE HTML>
+<html>
+<head>
+	<title>OR: API Template</title>
+    <link rel="stylesheet" type="text/css" href="/index.css"><!-- share styles with home page -->
+	<script src="index.js"></script>
+</head>
+
+<body onload="poll(500)" ><!-- 500 mSec is the default and may be omitted -->
+  <div id=menu><a href="/">Back to Menu</a></div>
+  <h1>Open Rails - API Template</h1>
+
+    <b>Sample Boolean:</b>
+    <div id="boolData"></div><br />
+    <b>Sample Integer:</b>
+    <div id="intData"></div><br />
+    <b>Sample Float:</b>
+    <div id="floatData"></div><br />
+    <b>Sample String:</b>
+    <div id="strData"></div><br />
+    <b>Sample Date:</b>
+    <div id="dateData"></div><br />
+
+    <b>Sample Custom Object:</b><br />
+    <b>- String:</b>
+    <div id="customObjectStr"></div>
+    <b>- Integer:</b>
+    <div id="customObjectInt"></div><br />
+
+    <b>Sample String Array:</b>
+    <div id="arrayItem0"></div>
+    <div id="arrayItem1"></div>
+    <div id="arrayItem2"></div>
+    <div id="arrayItem3"></div>
+    <div id="arrayItem4"></div><br />
+
+    <b>Sample List of Integers:</b>
+    <div id="intList0"></div>
+    <div id="intList1"></div>
+    <div id="intList2"></div>
+
+</body>
+</html>
+
+
+
