+++ Source/RunActivity/Viewer3D/WebServices/Web/API/TrackMonitor/index.html
@@ -1,14 +1,15 @@
 <!DOCTYPE HTML>
 <html>
 <head>
-	<title>OR: TrainInfo API</title>
+	<title>OR: Track Monitor API</title>
+	<link rel="shortcut icon" type="image/png" href="/or_logo.png"><!-- This icon appears in the page's tab and its bookmark. -->
     <link rel="stylesheet" type="text/css" href="/index.css"><!-- share styles with home page -->
 	<script src="index.js"></script>
 </head>
 
 <body onload="poll(500)" ><!-- 500 mSec is the default and may be omitted -->
   <div id=menu><a href="/">Back to Menu</a></div>
-  <h1>Open Rails Web Server - Api TrainInfo</h1>
+  <h1>Open Rails - Track Monitor (F4)</h1>
 
   <div>
     <table style="padding-left: 50px;">
